Fix dataset feature save
